python/requirements.txt
Flask==2.3.2
gunicorn==21.2.0
flask-cors==3.0.10
anthropic==0.34.1
python-dotenv==1.0.1